Update rem.js
Added removeBrowserSpecificQueries function to remove @-webkit(…) and @-moz(…) queries. If a @-webkit or @-moz query is present before a media query, css.match(/\}\s*\}/) will not match the end of the correct media query but the end of the browser specific query instead (which is never matched and never removed otherwise). The function is called in removeMediaQueries for browser that don't support mediaQueries.